+++ src/pudl/transform/__init__.py
@@ -71,9 +71,11 @@ from . import (
     eia_bulk_elec,
     eiaaeo,
     epacems,
+    ferc,
     ferc1,
     ferc714,
     gridpathratoolkit,
     nrelatb,
     params,
+    vceregen,
 )